+++ examples/pydantic/tutorial_2.py
@@ -45,10 +45,10 @@ async def run():
     tourpy = await Tournament_Pydantic_List.from_queryset(Tournament.all())
 
     # As Python dict with Python objects (e.g. datetime)
-    # Note that the root element is '__root__' that contains the root element.
-    print(tourpy.dict())
+    # Note that the root element is 'root' that contains the root element.
+    print(tourpy.model_dump())
     # As serialised JSON (e.g. datetime is ISO8601 string representation)
-    print(tourpy.json(indent=4))
+    print(tourpy.model_dump_json(indent=4))
 
 
 if __name__ == "__main__":